finished resetting CameraOpen bottom photos.
I was able to change the photolist from location.js by route.param ing the state and the photolist
to the location page. Added a simple function L97 that gets called on the async function storeData
that gets called when user finishes the snaq and relocated back to the home page. I decided to make
it like this so that if we implement a back button on the location page, the state of the camera
photolist will be maintained.

:bug:NOTE:bug:

I was getting an error when first testing but found it to be hard to replicate. It had something to
do with route.param ing state and manipulating state through those means. I have also tried a full
code refactor using useContext() and useContextProvider, but unfortunately, i couldnt get it to work.
There is also another workaround to what I have done, which also requires a lot of code refactoring.
It involves creating a new parent component that houses the CO screen and the Location screen, so
we could have a state(s) maintained by a parent. But i also decided against this since almost all
the states/vars are useless in Location.