Added marker state to allow user to move marker freely on click
Changed the marker in Mapview to check markerData. If null, marker sets to current location. On map click, it moves the marker to clicked area or POI(point of interest).

THINGS TO ADD:electric_plug:

Still need to find a way to update if marker was searched or not to allow only one marker showing on map. Currently, on Search creates a new marker that is immovable and still allows the markerData to move freely causing some UX confusion.
I have tried : { flag ? <searchedlocationMarker /> : <markerDataMarker /> } which would work, but finding it hard to setFlag to false on mapView. Also need to add a callout to the marker for better UX, but problem for a later date.

:bug: BUGS :bug:

current location button now creates another marker so that needs to be redisgned aswell to check whether or not there is a marker in markerData, and to setMarkerData back to latitude: lat, longitude: lng.